Add files via upload
Updated the "Teams" and "Our Work" pages. Added a mobile navigation menu which led me to altering the version of 'react-router-dom' to allow use of the '.active' class, which wasn't working in the most up to date version. Reverted back to RRD version 6.3.0. Also adjusted cubic-beziers for hero text (to prevent fall back) as well as the timing function for ScrollToTop. Also, MAKE SURE TO USE HASH ROUTER INSTEAD OF BROWSER ROUTER. I ran into a build bug.