Add files via upload
Added "Privacy Policy" and "Contact" pages. Also added transition animations to the "Services" and "Our Work" pages. Removed the parallax transition on smaller devices. Also added a scroll navigation to provide navigation when further down a page. Lastly started to mess with the tabIndex of the links on the home page.