Add files via upload
Added parallax effect to page banners. Added hover effects to the footer links. Also added the footer to each page so that it transitions with the rest of the page, instead of being stuck in place on page change. Also corrected an issue with the auto email generation on the services card.