Add files via upload
Finalized Version 1 - Added a 'Home' button to the navigation (and fixed the choppy hover animation) Also changed the 'Team' page to 'About' (across whole site) Changed the contact page to follow a calendly link for now. Replaced the skew animations (except for services page which is to be altered in the near future). Built an image modal for the 'Our Work' page and updated the copy on each page's banner.